reducing corrhist size
before: info depth 30 seldepth 46 time 155304 nodes 198206858 nps 1276250 hashfull 1000 score cp 111 pv d2d4 e7e6 c2c4 g8f6 g1f3 d7d5 b1c3 c7c5 c1g5 c5d4 f3d4 d5c4 e2e3 c8d7 f1c4 f8e7 e1g1 b8c6 d4f3 e8g8 d1e2 d8a5 g5h4 f8d8 a2a3 d7e8 b2b4 a5c7 h4g3 e7d6
after: info depth 30 seldepth 46 time 152409 nodes 198206858 nps 1300493 hashfull 1000 score cp 111 pv d2d4 e7e6 c2c4 g8f6 g1f3 d7d5 b1c3 c7c5 c1g5 c5d4 f3d4 d5c4 e2e3 c8d7 f1c4 f8e7 e1g1 b8c6 d4f3 e8g8 d1e2 d8a5 g5h4 f8d8 a2a3 d7e8 b2b4 a5c7 h4g3 e7d6

+++ thread.go
@@ -12,7 +12,7 @@
 //   --------
 //   One SearchState is created per thread slot and reused across
 //   moves of the same game.  Heuristic tables (history, correction)
-//   are NOT reset betwee\n moves so ordering signal accumulates.
+//   are NOT reset between moves so ordering signal accumulates.
 //   clearHistory() is called only on ucinewgame.
 //   resetForSearch() resets progress counters and per-ply context
 //   before each think() call.
@@ -55,12 +55,12 @@ type SearchState struct {
 	quietsMade   [maxPly][maxMoves]int // quiet moves tried so far at a current ply
 
 	// ---- Heuristic tables (persist across moves of the same game) ----
-	histTable       [2][64][64]int          // butterfly history [side][from][to]
-	contHistMain    [2][6][64][2][6][64]int // continuation history
-	killerMoves     [maxPly][2]int          // two killers per ply
-	moveBuffers     [maxPly]MovePicker      // pre-allocated pickers (one per ply)
-	pawnCorrHist        [2][corrHistSize]int    // pawn correction history
-	nonPawnCorrHist [2][2][corrHistSize]int // non-pawn correction history
+	histTable       [2][64][64]int            // butterfly history [side][from][to]
+	contHistMain    [2][6][64][2][6][64]int   // continuation history
+	killerMoves     [maxPly][2]int            // two killers per ply
+	moveBuffers     [maxPly]MovePicker        // pre-allocated pickers (one per ply)
+	pawnCorrHist    [2][corrHistSize]int16    // pawn correction history
+	nonPawnCorrHist [2][2][corrHistSize]int16 // non-pawn correction history
 }
 
 // State destroyed by makeMove.
@@ -93,8 +93,8 @@ func (ss *SearchState) clearHistory() {
 	ss.histTable = [2][64][64]int{}
 	ss.contHistMain = [2][6][64][2][6][64]int{}
 	ss.killerMoves = [maxPly][2]int{}
-	ss.pawnCorrHist = [2][corrHistSize]int{}
-	ss.nonPawnCorrHist = [2][2][corrHistSize]int{}
+	ss.pawnCorrHist = [2][corrHistSize]int16{}
+	ss.nonPawnCorrHist = [2][2][corrHistSize]int16{}
 }
 
 // resetForSearch prepares ss for a new search without losing
@@ -115,9 +115,9 @@ func (ss *SearchState) staticEval(p *Pos, ply int) int {
 	return evaluate(p, &ss.accStack[ply])
 }
 
-// doMove() is a wrapper for makemove to simplify search code
-// by hiding stacks. It also returns pointer to updateStack,
-// required by nnue accumulator.
+// doMove makes a move, stores all undo and NNUE update data,
+// and prefetches the resulting position's TT bucket.
+// It returns pointer to updateStack, required by nnue accumulator.
 func (ss *SearchState) doMove(p *Pos, ply, move int) *Update {
 	u := &ss.updateStack[ply]
 	r := &ss.revertStack[ply]
